requirements.txt
Flask
Flask-SQLAlchemy
Flask-Limiter 
flask-talisman
python-dotenv
qrcode[pil]
Pillow
gunicorn
bleach
